Neutral and Bad Tiles Finished
As a note, each image that says: "NeutralTileX.png" (X is the number), that is supposed to represent all of the neutral tiles that won't be removed. The one's that may be removed is all the "BadTileX.png" (X again is the number). I have in total 7 neutral tiles and 3 bad tiles. Apologies that they are very random and lower effort. I still had the animation, character design, background art, and potentially the music still left, so I cut corners (pun intended).